rename submodule `fee` with `percentage_with_cap`
