+++ zk-sdk/src/sigma_proofs/percentage_with_cap.rs
